Refactored deck listing logic into a component
